Bug in runs-up test solved. Anotated results for SplitMix tests.
In tests/test.c, a buffer that were too small was causing a
segfault. This was happening because to support compilers like Visual
Studio that do not support declaring arrays with variable size, we had
to substitute those variable with constants. And an error was commited
while figuring the correct array size in that case.

We also inlined function splitmix_next. This could be neccessary to
provide a more fair comparison between the execution times of
SplitMix64 and other algorithms. Other algorithms do not use
additional function calls where computing the next pseudo-random
number, so we should not require this for SplitMix.

Finally, we updated the tables with SplitMix results in the quality
tests and benchmark in OpenBSD.