Adding equidistribution test for 32 bits.
The current equidistribution test is too naive. It is a bit
equidistribution. We should check for 32-bit equidistribution. It is
probably the gretest power of 2 such that a user can generate in a
single execution of a program. As 2^32 random numbers could be
generated in 42 seconds if you can generate one per nanossecond. But
at the same rate, we need some centuries to generate 2^64 random
numbers.