Added a debug mode to the API.
Now you can run the API in debug mode defining the macro
"W_DEBUG_RNG". This will make the API print which random number
generator algorithm is being initialized and with what seed.

A bug was also corrected. The API was not correctly identifying when a
user choose the SPlitMix algorithm.